Estimate + Load Tasks Implementation
Completed an implementation of memory maintenance and N-back tasks in the advisorChoice class. Code in Index is commented out to add these trials to the experiment. Currently pending on adding a method of setting load trials structure in the Index file. Also involved adding method to utils for generating a random number within an upper and lower bound. There is now also an example advisor estimate task with some example MBA attributes similar to Dietvorst in the mbaVar.js file. Will need to look, again, at allowing users to set their experiment in the Index file.